refatoração da deature v02 #17 #18
