+++ .idea/misc.xml
@@ -4,7 +4,16 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="app/src/main/res/layout/activity_detail.xml" value="0.321875" />
+        <entry key="app/src/main/res/layout/activity_home.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.14895833333333333" />
+        <entry key="app/src/main/res/layout/activity_none.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_detail.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_globe.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_like.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_none.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.35104166666666664" />
+        <entry key="app/src/main/res/layout/fragment_search.xml" value="0.35104166666666664" />
         <entry key="app/src/main/res/layout/movie_item.xml" value="0.33" />
       </map>
     </option>
