+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.cancio.filmin" >
+    package="dev.cancio.popcorn">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -10,13 +10,16 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Filmin" >
+        android:theme="@style/Theme.Filmin">
         <activity
-            android:name=".ui.activity.DetailActivity"
+            android:name="dev.cancio.popcorn.ui.activity.HomeActivity"
             android:exported="true" />
         <activity
-            android:name=".ui.activity.HomeActivity"
-            android:exported="true" >
+            android:name="dev.cancio.popcorn.ui.activity.DetailActivity"
+            android:exported="true" />
+        <activity
+            android:name="dev.cancio.popcorn.ui.activity.MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
