--- app/src/main/java/dev/cancio/filmin/MyApplication.kt
@@ -1,9 +1,0 @@
-package dev.cancio.filmin
-
-import android.app.Application
-import dev.cancio.filmin.di.DaggerMainComponent
-
-class MyApplication: Application() {
-
-    val appComponent = DaggerMainComponent.builder().build()
-}
