--- app/src/main/java/dev/cancio/filmin/base/BaseView.kt
@@ -1,9 +1,0 @@
-package dev.cancio.filmin.base
-
-import androidx.viewbinding.ViewBinding
-
-interface BaseView{
-    val binding: ViewBinding
-
-    fun bindViews()
-}
