--- app/src/main/java/dev/cancio/filmin/presenter/HomePresenter.kt
@@ -1,41 +1,0 @@
-package dev.cancio.filmin.presenter
-
-import dev.cancio.filmin.base.BasePresenter
-import dev.cancio.filmin.base.BaseView
-import dev.cancio.filmin.data.model.Movie
-import dev.cancio.filmin.data.model.MoviePagination
-import dev.cancio.filmin.data.repository.MovieRepository
-import javax.inject.Inject
-
-class HomePresenter @Inject constructor(
-    private val view:View,
-    private val movieRepository : MovieRepository
-): BasePresenter<HomePresenter.View>(view) {
-
-    fun getMoviesList(){
-        launch {
-            val response = movieRepository.getMoviesList()
-            if(response.isSuccessful){
-                val moviePagination = response.body()
-                moviePagination?.apply {
-                    view.inflateRecyclerView(results)
-                }
-            }else{
-                view.onError()
-                onCallMock()
-            }
-        }
-    }
-
-    private fun onCallMock(){
-        val fileName = "mock_movie_pagination.json"
-        val mock = readJsonMock<MoviePagination>(fileName,MoviePagination::class.java)
-        val results = mock.results
-        view.inflateRecyclerView(results)
-    }
-
-    interface View: BaseView{
-        fun inflateRecyclerView(movieList: List<Movie>)
-        fun onError()
-    }
-}
