+++ app/src/main/java/dev/cancio/popcorn/MyApplication.kt
@@ -1,0 +1,9 @@
+package dev.cancio.popcorn
+
+import android.app.Application
+import dev.cancio.popcorn.di.DaggerMainComponent
+
+class MyApplication: Application() {
+
+    val appComponent = DaggerMainComponent.builder().build()
+}
