+++ app/src/main/java/dev/cancio/popcorn/base/BaseView.kt
@@ -1,0 +1,9 @@
+package dev.cancio.popcorn.base
+
+import androidx.viewbinding.ViewBinding
+
+interface BaseView{
+    val binding: ViewBinding
+
+    fun bindViews()
+}
