+++ app/src/main/java/dev/cancio/popcorn/di/MainModule.kt
@@ -1,0 +1,15 @@
+package dev.cancio.popcorn.di
+
+import dagger.Module
+import dagger.Provides
+import dev.cancio.popcorn.data.service.MovieService
+import retrofit2.Retrofit
+
+@Module
+class MainModule {
+
+    @Provides
+    fun provideMovie(retrofit: Retrofit): MovieService {
+        return retrofit.create(MovieService::class.java)
+    }
+}
