+++ app/src/main/java/dev/cancio/popcorn/ui/adapter/MovieItemAdapter.kt
@@ -1,0 +1,28 @@
+package dev.cancio.popcorn.ui.adapter
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import dev.cancio.popcorn.data.model.Movie
+import dev.cancio.popcorn.databinding.MovieItemBinding
+import dev.cancio.popcorn.ui.vh.MovieItemViewHolder
+
+class MovieItemAdapter(
+    private val context: Context,
+    private val movieList: List<Movie>
+) : RecyclerView.Adapter<MovieItemViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieItemViewHolder {
+        val binding = MovieItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+
+        return MovieItemViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: MovieItemViewHolder, position: Int) {
+        val movie = movieList[position]
+        holder.bind(movie)
+    }
+
+    override fun getItemCount(): Int = movieList.size
+}
