+++ app/src/main/res/layout/activity_home.xml
@@ -23,7 +23,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/green_03"
-        app:itemIconTint="@color/grey"
+        app:itemIconTint="@color/grey_1"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
