+++ app/src/main/res/layout/movie_item.xml
@@ -24,7 +24,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Cruelly"
-        android:textColor="@color/grey"
+        android:textColor="@color/grey_1"
         android:layout_margin="8dp"
         app:layout_constraintStart_toEndOf="@+id/imageView_poster"
         app:layout_constraintTop_toTopOf="parent" />
@@ -55,7 +55,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Craig Gillespie"
-        android:textColor="@color/grey"
+        android:textColor="@color/grey_1"
         app:layout_constraintBottom_toBottomOf="@+id/imageView_poster"
         app:layout_constraintStart_toStartOf="@+id/imageView_star" />
