+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,15 @@
 <resources>
-    <string name="app_name">filmin</string>
+    <string name="app_name" translatable="false">popcorn</string>
     <string name="star_rating">star rating</string>
     <string name="poster_description">poster about the movie</string>
+    <string name="fragment_title_like">Likes</string>
+    <string name="fragment_title_search">Search</string>
+    <string name="fragment_title_globe">Globe</string>
+    <string name="fragment_title_home">Featured</string>
+    <string name="fragment_title_profile">Profile</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="none_page">This page doesn\'t exist (yet)</string>
+    <string name="fragment_detail">Details</string>
+    <string name="app_title" translatable="false">PopCorn</string>
     <string name="error_message">"The connection has failed"</string>
-    <string name="bottom_bar_item_home">Home</string>
 </resources>
