+++ app/src/main/res/values/themes.xml
@@ -1,13 +1,13 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.Filmin" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.Filmin" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/green_03</item>
+        <item name="colorPrimary">@color/green_01</item>
         <item name="colorPrimaryVariant">@color/green_02</item>
-        <item name="colorOnPrimary">@color/grey</item>
+        <item name="colorOnPrimary">@color/grey_1</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/green_02</item>
-        <item name="colorSecondaryVariant">@color/green_01</item>
+        <item name="colorSecondaryVariant">@color/green_03</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
